Move files in the phone bill original project to the "phonebill" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
