+++ projects-parent/originals-parent/phonebill/src/it/java/edu/pdx/cs410J/phonebill/Project1IT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.phonebill;
 
 import edu.pdx.cs410J.InvokeMainTestCase;
 import org.junit.Test;
